Separate out test packet generation code.
